refactor: 각 기물에서 중복되는 로직을 Piece에서 수행하도록 수정, 모든 기물 삭제
