--- src/main/java/domain/piece/Cannon.java
@@ -1,7 +1,0 @@
-package domain.piece;
-
-public class Cannon extends Piece {
-    public Cannon(PieceInfo pieceInfo) {
-        super(pieceInfo);
-    }
-}
