--- src/main/java/domain/piece/Chariot.java
@@ -1,7 +1,0 @@
-package domain.piece;
-
-public class Chariot extends Piece {
-    public Chariot(PieceInfo pieceInfo) {
-        super(pieceInfo);
-    }
-}
