--- src/main/java/domain/piece/Elephant.java
@@ -1,7 +1,0 @@
-package domain.piece;
-
-public class Elephant extends Piece {
-    public Elephant(PieceInfo pieceInfo) {
-        super(pieceInfo);
-    }
-}
