--- src/main/java/domain/piece/General.java
@@ -1,7 +1,0 @@
-package domain.piece;
-
-public class General extends Piece {
-    public General(PieceInfo pieceInfo) {
-        super(pieceInfo);
-    }
-}
