--- src/main/java/domain/piece/Guard.java
@@ -1,7 +1,0 @@
-package domain.piece;
-
-public class Guard extends Piece {
-    public Guard(PieceInfo pieceInfo) {
-        super(pieceInfo);
-    }
-}
