--- src/main/java/domain/piece/Horse.java
@@ -1,7 +1,0 @@
-package domain.piece;
-
-public class Horse extends Piece {
-    public Horse(PieceInfo pieceInfo) {
-        super(pieceInfo);
-    }
-}
