+++ src/test/java/domain/piece/PieceTest.java
@@ -11,16 +11,16 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class SoldierTest {
+public class PieceTest {
     @ParameterizedTest
-    @DisplayName("목적지까지의 경로를 정확히 계산한다.")
-    @MethodSource("expectedPaths")
-    void pathTest(Position from, Position to, List<Position> paths) {
-        Soldier soldier = new Soldier(new PieceInfo(PieceType.SOLDIER, Country.CHO));
+    @DisplayName("졸병의 목적지까지의 경로를 정확히 계산한다.")
+    @MethodSource("expectedSoldierPaths")
+    void soldierPathTest(Position from, Position to, List<Position> paths) {
+        Piece soldier = new Piece(new PieceInfo(PieceType.SOLDIER, Country.CHO));
         assertThat(soldier.path(from, to)).isEqualTo(paths);
     }
 
-    static Stream<Arguments> expectedPaths() {
+    static Stream<Arguments> expectedSoldierPaths() {
         return Stream.of(
                 Arguments.arguments(new Position(1, 1), new Position(1, 2),
                         List.of(new Position(1, 1), new Position(1, 2))),
